+++ src/main/java/janggi/factory/PieceFactory.java
@@ -5,10 +5,10 @@ import janggi.domain.move.Position;
 import janggi.domain.Side;
 import janggi.domain.piece.behavior.Cannon;
 import janggi.domain.piece.behavior.Chariot;
-import janggi.domain.piece.behavior.Elephant;
+import janggi.domain.piece.behavior.crossmove.Elephant;
 import janggi.domain.piece.behavior.General;
 import janggi.domain.piece.behavior.Guard;
-import janggi.domain.piece.behavior.Horse;
+import janggi.domain.piece.behavior.crossmove.Horse;
 import janggi.domain.piece.PieceBehavior;
 import janggi.domain.piece.behavior.Soldier;
 import java.util.HashMap;
